Fix bug in can_be_true for expr objects with no Variable nodes
If `expression` does not have any Variable nodes, the underlying call
to get_variables/GetDeclarations returns nothing, and the related_to()
call returns an essentially empty set of constraints which will always
be true. This ignores the constraints completely, which is incorrect.

I believe the related_to() here is an optimization, which can be removed
for correctness.