Explicitly override __hash__ to use object.__hash__ (#835)
In python3 there are changes to how you are supposed to handle
overrides of __eq__, as it relates to __hash__
(http://python3porting.com/preparing.html?highlight=hash#implementing-hash)

In python2 we changed __eq__, but did nothing for __hash__, so __hash__
kept using the default one provided by object. In python3 it automatically
sets __hash__ to None, so we need to explicitly emulate this.