Eth updates (#1195)
* Add offset param to interface for extract a uint of variable with

* add leniency for g

* Rename attr

* other change

* bugs

* Automatically truncate sym vars that are too big

This is opinionated. Users use m.make_symbolic_value() to pass symbolic
arguments to contract functions. this creates a 256 bitvec by default.
if they do this for an argument that is not 256 bits, an `address` for
example, this would previously cause an AssertionError in ZEXTEND
because we're trying to ZEXTEND down to a smaller size (160 bits).

Manticore has all the information it needs to do the right thing here,
and automatically truncate the symbolic variable, so this implements
that.

* Fix logic bug in _readBE

* make padding default False because it's never used?

* h tx

* clean comment

* clearer comment

* fix bug, size should always be nbytes

* move type forgiveness up to the serialize() level

* fix logic bug - correctly loop

* update callsites correctly

* better err msg

* fix

* add tests for str args to ctor