Use the address of symbolic conditional instructions
When two symbolic conditional instructions follow sequentially, the two paths of
the first instruction will be explored. But when the exploration reaches the
second instruction, because the at_symbolic_conditional flag has already been
set by the first one, only one of the two possible paths will be explored.

This commit fixes this issue by replacing the boolean flag by the address of the
instruction, which ensure the flag has been set for the current instruction.