Fix emulation being performed in the wrong mode
After the merge of the PR referenced at the end of this message into the master
branch of Unicorn in July 2016, the emulation mode is set according to the
least significant of bit of the program counter when it is updated.

When an instruction is not implemented at Manticore's level and emulation is
used, the PC value passed to the emu_start() function will trigger a switch of
the current mode, setting it back to ARM mode instead of Thumb mode. This commit
fixes this issue by ensuring that the least significant bit of the PC is set
when Thumb mode emulation is performed.

See https://github.com/unicorn-engine/unicorn/pull/592