Fix wrong mode being pickled for the initial state only
If the entrypoint of an program is located on a Thumb mode instruction (if the
least significant bit of the entrypoint is set), the wrong mode will be saved in
the pickled initial state because the mode switching is only happening later.

This commit fixes this issue by switching mode before the pickling, and
unsetting the LSbit of the PC to avoid reading the wrong instruction bytes.