Added ADR, ADDW and SUBW instructions semantics
These 3 instructions can be used to load PC-relative addresses. The ADDW and
SUBW instructions can also be used with other source registers. When the PC
register is used as the source, the ARM documentation specifies that the offset
is added or subtracted from ALIGN(pc, 4).

See also the ARM Architecture Reference Manual Thumb-2 Supplement.