Allow file offsets not aligned on page size (#1242)
* Allow file offsets not aligned on page size

When mapping into memory uncommon ELF executables having their segments file
offsets not aligned on the system page size, the call to mmap_function() will
fail silently. It will actually return 0xffffffff, which isn't checked anywhere
and will only be detected when trying to read/write/free the memory area.

Because this ELF files triggering this issue are programmatically crafted from
binaries into a proprietary format, it won't be happening for all other
executables. But Manticore should support binaries as generic as possible.

This commit allows mapping into memory files having offsets not aligned on the
system page size by simply aligning the file offset when mmap'ing, and returning
a pointer shifted from the necessary correction. When munmap'ing the memory, the
address is aligned again to recovery the initial one given when mmap'ing.

* Fixed the size passed to mmap and added comments

* Fixed the size passed to unmap

* Added unit test for unaligned mappings

* Update comment

* Added new asserts, reformatted comments

* Checking if the CI fails because of the removed cast