Added CBZ, TBB and TBH instructions semantics (#1243)
* Added CBZ, TBB and TBH instructions semantics

CBZ was not implemented while CBNZ was. As far as Manticore is concerned, CBZ is
the same as CBNZ, but with two possible destination addresses exchanged.

TBB and TBH instructions are often used for switches. According to the ARM
documentation, the base register contains the address of a table of
bytes/half-words, and the index register contains an index into this table. The
selected value is then shifted left by one and zero-extended to 32-bits before
being added to the current PC. If the PC register is used as the base, the table
is then located immediately after the instruction.

See also ARM Architecture Reference Manual Thumb-2 Supplement.

* Added docstring and comments to instructions

* Added unit tests for the cbz/cbnz/tbb/tbh instructions

* Taking into account the review