Remove kwargs from ManticoreEVM
Removing `kwargs` from `ManticoreEVM` as it is not needed here.

It is because it is passed to `Manticore` which uses `kwargs` only in
the case when the `path_or_state` is a path. The `ManticoreEVM` passes
this argument as already created `State` so the `kwargs` are never used
for `ManticoreEVM`.

This will also prevent users from using old API that has been removed
e.g. `ManticoreEVM(verbosity=5)`.