evm: fix _check_jumpdest when run with detectors
When ManticoreEVM is run with detectors, the self.pc in EVM is a
Constant (as we need to taint it).

Because of that, the check:
```
            if pc not in self._valid_jumpdests:
                raise InvalidOpcode()
```
failed - i.e. we raised an InvalidOpcode because we checked if the list
contains the `Constant` object and not its value.

This commit fixes this issue.