Fixes ConstraintSet.new_bitvec size check
Previously the `new_bitvec` checked for `size == 1 or size % 8 == 0`.

This check allowed to create a bitvec with negative size, as in `cs.new_bitvec(-8)`.

It was also too constrained, as we might have cases where we want to constraint ourselves to e.g. 5 bits out of 8 bits of a byte). Sure, this can be done via setting a constrain over an 8-bit (1 Byte) bitvector, but it might be not the most efficient solution(?).