Improve WASM Memory Ergonomics (#1589)
* Prevent most modules from accessing memory._data directly

* Fix broken assert

Cheap way of getting the memory size, but it'll work

* Lazily create memory pages as-needed

* Add (disabled) support for name sections

The WASM importer module we're currently relying on breaks if you try to import the spectest binaries with name sections enabled. Dumb, I know.

* Fix misnamed events

* Add imports to postfilled import names

* Fix did->will set global

* Fix memory callbacks and add tests

* Update structure.py

* mypy

* Revert "Update structure.py"

This reverts commit bedb72265979a6c33f41b3edd9b25ae0f9a26071.

* Print module names and function returns

* Fix docstrings for function names

* Blacken

* Fix type declaration

* Add SerializeState exception

Raise `SerializeState("checkpoint.pkl")` to dump a state to the disk, which you can later load and resume execution from. Needs unit tests.

* PAGESIZE constant

* Make decoding names a config value

* Attempt to add type hints to memory io operations

They won't hold up to anything symbolic so we'll need to keep an eye on them