Rework syscall invocation for proper behavior under typeguard
When a syscall is made from an emulated binary, Manticore uses the
syscall number to look up the appropriate Python method that models that
syscall, and then uses Python introspection to massage arguments to that
syscall model function as deemed appropriate.

Previously, the deprecated `inspect.getfullargspec` method was used to
determine the number of arguments to the model function, and whether or
not it takes varargs.

However, in addition to being deprecated, `inspect.getfullargspec` does
not peer through wrapper functions (which can be introduced, for
example, through decorator use).  This inability to peer through wrapper
functions was causing Manticore's syscall emulation to break when the
Typeguard plugin for pytest is used -- that plugin implicitly adds a
`@typeguard.check_types` decorator on every function & method in the
`manticore` package.

Now, we use the non-deprecated `inspect.signature` API to get the needed
information.  This API _does_ peer through wrapper functions, and it
also allows us to get rid of conditional logic around the inclusion of
`self` in bound methods.