Remove some wrong type hints from AbstractCpu
It turns out that the type of `data` in a couple methods is pretty flexible,
and our tests use more than just `str`, `bytes`, and `List` types for that
parameter.

It is not quite clear what requirements there actually are on `data`, and so at
least for now, get rid of the type hint there.