Add a helper for working with Python abstract base classes
This helper lets you declare on a class, via a decorator, that the class
is expected to have no unimplemented abstract methods.  If this is not
true at runtime, a TypeError is thrown.

Without this decorator, you will only discover that a class has
unimplemented abstract methods when you try to _instantiate_ that class.
This may be much, much later than at class creation time, or perhaps not
at all, depending on your code coverage.