Kill Manticore if _any_ state encounters unrecoverable exception
This is useful for CHESS because we of how Manticore tasks are killed
through signals. Previously, a signal/exception is caught by the current
active state and then printed. This is an issue because if there is more
than one state, Manticore doesn't actually die when we want.

This change now kills Manticore when _any_ state encounters an
unrecoverable exception or assertion and operates as we expect upon a
death signal.

In the future, when upstreaming this branhc, we probably want to make
this opt-in to preserve previous functionality expectations.