Copy/Snapshot a RegisterFile (#2489)
This is useful for referencing the state of the registers at a certain point in time.

Internally, this is used to check the concrete values of registers. Symbolic values should also work but require more testing and care. Symbolic values containing memory references are undefined and are likely not accurate if execution occurs after the snapshot.