Add walrus operator to shorten code
PLEASE NOTE: This will ONLY function on Python 3.8 or above. If you do not intend running this within those versions, please discard my pull request. Thank you!
This will shorten the code by adding the walrus operator, which can set variables inside of expressions. If there happen to be any other sections that could use this, it could be very helpful in shortening more code in this repository