Solver Improvements (#2502)
In the course of my investigation, I found that the SMT2 definitions were being converted to text properly, but not making it all the way to the solver. This seems to be an issue with Python's unbuffered pipe implementation. I'm not sure what the exact issue is, but switching to buffered I/O and adding explicit calls to flush has fixed the tests.

* Fix invalid escape characters

* Add debug logs for solver errors

* Optimize imports

* Split buffer on newlines

We're going to have to take this out before merging but hopefully it'll help figure out what's going wrong

* Clear the debug buffer upon reset

* Make sure we flush data sent to the solver

* Fix typo

* Try buffering I/O with the solver

* Move _clear_buffers out of loop

* Remove debugging code